# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Crop Disease Detector
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PredictionSeverity =
  (typeof PredictionSeverity)[keyof typeof PredictionSeverity];

export const PredictionSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Prediction {
  id: number;
  cropType: string;
  diseaseName: string;
  confidence: number;
  severity: PredictionSeverity;
  treatment: string;
  imageUrl?: string | null;
  isHealthy: boolean;
  createdAt: string;
}

export interface CreatePredictionBody {
  cropType: string;
  imageBase64?: string | null;
  imageUrl?: string | null;
}

export type DiseaseSeverity =
  (typeof DiseaseSeverity)[keyof typeof DiseaseSeverity];

export const DiseaseSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Disease {
  id: number;
  name: string;
  cropType: string;
  description: string;
  symptoms: string;
  treatment: string;
  severity: DiseaseSeverity;
  imageUrl?: string | null;
}

export interface StatsSummary {
  totalScans: number;
  diseasesDetected: number;
  healthyPlants: number;
  averageConfidence: number;
  mostCommonDisease?: string | null;
}

export interface DiseaseCount {
  diseaseName: string;
  count: number;
  percentage: number;
}

export interface ErrorResponse {
  error: string;
}