# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Crop Disease Detector
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: predictions
    description: Disease prediction operations
  - name: diseases
    description: Disease information
  - name: stats
    description: Statistics and analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /predictions:
    get:
      operationId: listPredictions
      tags: [predictions]
      summary: List all predictions
      description: Returns all disease predictions made
      responses:
        "200":
          description: List of predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"
    post:
      operationId: createPrediction
      tags: [predictions]
      summary: Create a disease prediction
      description: Submit an image and receive disease prediction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePredictionBody"
      responses:
        "201":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"

  /predictions/{id}:
    get:
      operationId: getPrediction
      tags: [predictions]
      summary: Get a prediction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prediction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /diseases:
    get:
      operationId: listDiseases
      tags: [diseases]
      summary: List all known diseases
      responses:
        "200":
          description: List of diseases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Disease"

  /diseases/{id}:
    get:
      operationId: getDisease
      tags: [diseases]
      summary: Get a disease by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disease details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disease"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get dashboard statistics summary
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/disease-distribution:
    get:
      operationId: getDiseaseDistribution
      tags: [stats]
      summary: Get distribution of detected diseases
      responses:
        "200":
          description: Disease distribution
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiseaseCount"

  /stats/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Get recent detection activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Prediction:
      type: object
      properties:
        id:
          type: integer
        cropType:
          type: string
        diseaseName:
          type: string
        confidence:
          type: number
          format: float
        severity:
          type: string
          enum: [low, medium, high]
        treatment:
          type: string
        imageUrl:
          type: string
          nullable: true
        isHealthy:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - cropType
        - diseaseName
        - confidence
        - severity
        - treatment
        - isHealthy
        - createdAt

    CreatePredictionBody:
      type: object
      properties:
        cropType:
          type: string
        imageBase64:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required:
        - cropType

    Disease:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cropType:
          type: string
        description:
          type: string
        symptoms:
          type: string
        treatment:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        imageUrl:
          type: string
          nullable: true
      required:
        - id
        - name
        - cropType
        - description
        - symptoms
        - treatment
        - severity

    StatsSummary:
      type: object
      properties:
        totalScans:
          type: integer
        diseasesDetected:
          type: integer
        healthyPlants:
          type: integer
        averageConfidence:
          type: number
          format: float
        mostCommonDisease:
          type: string
          nullable: true
      required:
        - totalScans
        - diseasesDetected
        - healthyPlants
        - averageConfidence

    DiseaseCount:
      type: object
      properties:
        diseaseName:
          type: string
        count:
          type: integer
        percentage:
          type: number
          format: float
      required:
        - diseaseName
        - count
        - percentage

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error