# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Crop Disease Detector
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all disease predictions made
 * @summary List all predictions
 */
export const ListPredictionsResponseItem = zod.object({
  id: zod.number(),
  cropType: zod.string(),
  diseaseName: zod.string(),
  confidence: zod.number(),
  severity: zod.enum(["low", "medium", "high"]),
  treatment: zod.string(),
  imageUrl: zod.string().nullish(),
  isHealthy: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListPredictionsResponse = zod.array(ListPredictionsResponseItem);

/**
 * Submit an image and receive disease prediction
 * @summary Create a disease prediction
 */
export const CreatePredictionBody = zod.object({
  cropType: zod.string(),
  imageBase64: zod.string().nullish(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Get a prediction by ID
 */
export const GetPredictionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPredictionResponse = zod.object({
  id: zod.number(),
  cropType: zod.string(),
  diseaseName: zod.string(),
  confidence: zod.number(),
  severity: zod.enum(["low", "medium", "high"]),
  treatment: zod.string(),
  imageUrl: zod.string().nullish(),
  isHealthy: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all known diseases
 */
export const ListDiseasesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  cropType: zod.string(),
  description: zod.string(),
  symptoms: zod.string(),
  treatment: zod.string(),
  severity: zod.enum(["low", "medium", "high"]),
  imageUrl: zod.string().nullish(),
});
export const ListDiseasesResponse = zod.array(ListDiseasesResponseItem);

/**
 * @summary Get a disease by ID
 */
export const GetDiseaseParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDiseaseResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  cropType: zod.string(),
  description: zod.string(),
  symptoms: zod.string(),
  treatment: zod.string(),
  severity: zod.enum(["low", "medium", "high"]),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Get dashboard statistics summary
 */
export const GetStatsSummaryResponse = zod.object({
  totalScans: zod.number(),
  diseasesDetected: zod.number(),
  healthyPlants: zod.number(),
  averageConfidence: zod.number(),
  mostCommonDisease: zod.string().nullish(),
});

/**
 * @summary Get distribution of detected diseases
 */
export const GetDiseaseDistributionResponseItem = zod.object({
  diseaseName: zod.string(),
  count: zod.number(),
  percentage: zod.number(),
});
export const GetDiseaseDistributionResponse = zod.array(
  GetDiseaseDistributionResponseItem,
);

/**
 * @summary Get recent detection activity
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  cropType: zod.string(),
  diseaseName: zod.string(),
  confidence: zod.number(),
  severity: zod.enum(["low", "medium", "high"]),
  treatment: zod.string(),
  imageUrl: zod.string().nullish(),
  isHealthy: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);